Before Apply new Updates Create a backup first dont tell me that i didnt warn you! Remember to revert Changes i send in patch to some of you Update all libs from this version
Update (Config files) General.ini, adminCommands.xml and Server.ini (Check what is changed)
Add new xml folder located in config/ both in login and game server
Delete config file with name GeoDriver.properties in game/config/
Updated some server starters

Whole core improvements made!
i will list some of the fixes
1. Fixed all pvp skill issues land in non flag players
2.Removed Point3d class thanks l2j
3.Fixed Clan Community
4.Added Mail board in community
5.Reworked database connection to be more stable!!
6.Reworked all loggers!
7.Fix some summons and siege errors
8.Added retail like transformation system
9.Reworked buylist system now is in xml except one sql is created for your customs


Enjoy Enterprize version of Reunion project
Thanks all for waiting!
But now reunion project is more stable than ever!

+++ L2J_ReunionProject_Core/java/l2r/gameserver/pathfinding/geonodes/GeoNodeLoc.java
@@ -50,7 +50,7 @@ public class GeoNodeLoc extends AbstractNodeLoc
 	}
 	
 	@Override
-	public int getZ()
+	public short getZ()
 	{
 		return _z;
 	}
